+++ EpikCord/__init__.py
@@ -10,6 +10,7 @@ from typing import *
 from urllib.parse import quote
 import io
 import os
+from .application import *
 
 CT = TypeVar('CT', bound='Colour')
 T = TypeVar('T')
@@ -1159,23 +1160,7 @@ class BaseComponent:
         self.settings["custom_id"] = custom_id
 
 
-class Application:
-    def __init__(self, data: dict):
-        self.id: str = data.get("id")
-        self.name: str = data.get("name")
-        self.icon: Optional[str] = data.get("icon")
-        self.description: str = data.get("description")
-        self.rpc_origins: Optional[list] = data.get("rpc_origins")
-        self.bot_public: bool = data.get("bot_public")
-        self.bot_require_code_grant: bool = data.get("bot_require_code_grant")
-        self.terms_of_service_url: Optional[str] = data.get("terms_of_service")
-        self.privacy_policy_url: Optional[str] = data.get("privacy_policy")
-        self.owner: Optional[PartialUser] = PartialUser(data.get("user")) if data.get("user") else None
-        self.summary: str = data.get("summary")
-        self.verify_key: str = data.get("verify_key")
-        self.team: Optional[Team] = Team(data.get("team")) if data.get("get") else None
-        self.cover_image: Optional[str] = data.get("cover_image")
-        self.flags: int = data.get("flags")
+
 
 class InvalidApplicationCommandType(Exception):
     ...
