+++ EpikCord/application.py
@@ -1,1 +1,21 @@
-#publish
+from typing import Optional
+from .__init__ import PartialUser, Team
+#temp import
+# we have to make sure there is no circular imports, someone help us!
+class Application:
+    def __init__(self, data: dict):
+        self.id: str = data.get("id")
+        self.name: str = data.get("name")
+        self.icon: Optional[str] = data.get("icon")
+        self.description: str = data.get("description")
+        self.rpc_origins: Optional[list] = data.get("rpc_origins")
+        self.bot_public: bool = data.get("bot_public")
+        self.bot_require_code_grant: bool = data.get("bot_require_code_grant")
+        self.terms_of_service_url: Optional[str] = data.get("terms_of_service")
+        self.privacy_policy_url: Optional[str] = data.get("privacy_policy")
+        self.owner: Optional[PartialUser] = PartialUser(data.get("user")) if data.get("user") else None
+        self.summary: str = data.get("summary")
+        self.verify_key: str = data.get("verify_key")
+        self.team: Optional[Team] = Team(data.get("team")) if data.get("get") else None
+        self.cover_image: Optional[str] = data.get("cover_image")
+        self.flags: int = data.get("flags")
